requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
openai==1.3.7
pillow==10.1.0
pytesseract==0.3.10
pypdf2==3.0.1
python-docx==1.1.0
websockets==12.0
aiofiles==23.2.1
boto3==1.34.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
